fine-tuning/requirements.txt
openai>=1.2.0
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0 